Cleaned up a few potential memory leaks and other warnings under LLVM 1.5
Added autorelease messages in several instances of [somearray mutableCopy]

Added [someObject release] messages to match [[someClass alloc] init] lines where they had been omitted.

Explicitly set assign for @property (..,assign) Class klass declarations as LLVM 1.5 complains if omitted.

Modified the #import SBJSON.h to be #import JSON.h as per new naming in JSON Framework.

Added parentheses around assignment within if statements to remove compiler warnings.

Removed implicit casts of NSUInteger to int in methods that use [someCollection count] methods. Also changed the type of several iterator variables for similar reasons.

+++ CoreTableController.h
@@ -27,10 +27,10 @@
 #pragma mark -
 #pragma mark Results methods
 - (Class) model;
-- (int) resultsSectionCount;
+- (NSUInteger) resultsSectionCount;
 - (BOOL) hasResults;
 - (id) resultsInfoForSection:(int)section;
-- (int) resultsCountForSection:(int)section;
+- (NSUInteger) resultsCountForSection:(NSUInteger)section;
 - (CoreResource*) resourceAtIndexPath:(NSIndexPath*)indexPath;
 - (NSString*) noResultsMessage;
 - (BOOL) hasNoResultsMessage;
